improve arrow handling at top of divs and doc
